Minor Refactor
